backend/requirements.txt
asgiref==3.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
dask==2022.5.0
distributed==2022.5.0
Django==3.1
django-cors-headers==3.11.0
djangorestframework==3.12.0
drf-yasg==1.20.0
fsspec==2022.3.0
HeapDict==1.0.1
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.1.1
locket==1.0.0
MarkupSafe==2.1.1
msgpack==1.0.3
packaging==21.3
partd==1.2.0
paypalrestsdk==1.13.1
Pillow==9.1.0
psutil==5.9.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tblib==1.7.0
toolz==0.11.2
tornado==6.1
uritemplate==4.1.1
urllib3==1.26.9
zict==2.2.0
